Refactor delete task module
